+++ src/main/java/chess/domain/piece/type/Bishop.java
@@ -1,10 +1,10 @@
 package chess.domain.piece.type;
 
-import chess.domain.Movement;
+import chess.domain.position.Movement;
 import chess.util.RouteCalculator;
 import chess.domain.piece.Color;
 import chess.domain.piece.Piece;
-import chess.domain.piece.Position;
+import chess.domain.position.Position;
 import java.util.Set;
 
 public class Bishop extends Piece {
