+++ src/main/java/chess/domain/piece/type/King.java
@@ -1,9 +1,9 @@
 package chess.domain.piece.type;
 
-import chess.domain.Movement;
+import chess.domain.position.Movement;
 import chess.domain.piece.Color;
 import chess.domain.piece.Piece;
-import chess.domain.piece.Position;
+import chess.domain.position.Position;
 import java.util.HashSet;
 import java.util.Set;
