+++ src/test/java/chess/domain/piece/type/KingTest.java
@@ -2,11 +2,11 @@ package chess.domain.piece.type;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
-import chess.domain.Movement;
+import chess.domain.position.Movement;
 import chess.domain.piece.Color;
-import chess.domain.piece.File;
-import chess.domain.piece.Position;
-import chess.domain.piece.Rank;
+import chess.domain.position.File;
+import chess.domain.position.Position;
+import chess.domain.position.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
